reworked the API
components are now more fine grained and customisable

there is new Scorer interface which can be supplied to the annotator and which handles scoring of the terms (either positive or negative).

there is new TermMatcher interface which can be supplied to the annotator and which handles matching the terms in the letter.